refactor: move Dispatcher to common
