+++ packages/extension/src/types/emitter.ts
@@ -17,8 +17,7 @@
  ***********************************************************************/
 
 /* eslint-disable @typescript-eslint/no-unsafe-function-type */
-
-import type { IDisposable } from './disposable.js';
+import type { IDisposable } from '/@common/model/disposable';
 
 export type DisposableGroup = { push(disposable: IDisposable): void } | { add(disposable: IDisposable): void };
